requirements.txt
httpx>=0.27.0
feedparser>=6.0.0
supabase>=2.0.0
upstash-qstash>=2.0.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
